main/requirements.txt
sqlalchemy
pytest

aiogram
python-dotenv
SQLAlchemy